Modify 102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int i = 0, a = 1, b = 2, nxtn;
+	long int i = 0, a = 1, b = 2, nxtn;
 
 	printf("%d, %d", a, b);
 	while (i < 48)
